+++ src/main/java/cryptogame/model/dao/user/UserDao.java
@@ -6,12 +6,13 @@ import cryptogame.model.dao.Dao;
 import cryptogame.model.models.UserModel;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 public interface UserDao extends Dao<UserModel>
 {
-    void purchaseCurrency(UserModel user, double amount , CurrencyContainer currency);
-    void sellCurrency(UserModel user, double amount, CurrencyContainer currency);
+    void purchaseCurrency(UserModel user, BigDecimal amount , CurrencyContainer currency);
+    void sellCurrency(UserModel user, BigDecimal amount, CurrencyContainer currency);
     Optional<UserModel> getByUsername(String username);
     Optional<UserModel> getByEmail(String email);
 }
